requirements.txt
streamlit
openai
elevenlabs>=1.0.0
python-dotenv
